+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Reworked xc3_model and xc3_shader database and material assignment types.
 * Changed shader assignments inferred from textures to merge XYZ channels for color and specular color.
 
+### Fixed
+* Fixed an issue where map models would use the wrong alpha testing texture in some cases.
+
 ## 0.22.0 - 2026-05-26
 ### Added
 * Added support for Xenoblade X DE for xc3_model wimdo export.